Move security-related dialogs to a common directory
