+++ res/css/views/dialogs/security/_AccessSecretStorageDialog.scss
