+++ res/css/views/dialogs/security/_CreateSecretStorageDialog.scss
