+++ res/css/views/dialogs/security/_KeyBackupFailedDialog.scss
