+++ res/css/views/dialogs/security/_RestoreKeyBackupDialog.scss
